+++ src/repository/LoginRepository.java
@@ -4,6 +4,6 @@ import model.Dusr;
 
 public interface LoginRepository {
 
-	Dusr autenticar(Short usuario, String senha);
+	Dusr autenticar(Integer usuario, String senha);
 
 }
